pyproject.toml
[project]
name = "magic-the-gathering-workspace"
version = "0.1.0"
description = "Magic: The Gathering toolkit - MCP server, TUI, and core library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "mtg-core",
    "mtg-spellbook",
    "mtg-mcp",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-textual-snapshot>=1.0.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
]

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
mtg-core = { workspace = true }
mtg-spellbook = { workspace = true }
mtg-mcp = { workspace = true }

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["packages/*/src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["mtg_core", "mtg_mcp_server", "mtg_spellbook"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = ["packages/mtg-core/src", "packages/mtg-spellbook/src", "packages/mtg-mcp-server/src"]
exclude = ["tests/", "tests_core/", "tests_spellbook/"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = ["tests.*", "mtg_spellbook.tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
warn_return_any = false
check_untyped_defs = false
disable_error_code = ["var-annotated", "comparison-overlap", "call-arg"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["packages/mtg-core/tests_core", "packages/mtg-spellbook/tests_spellbook"]
pythonpath = ["packages/mtg-core/src", "packages/mtg-spellbook/src"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[dependency-groups]
dev = [
    "pypdf>=6.4.2",
    "pytest>=9.0.2",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.8.0",
    "types-pyyaml>=6.0.12.20250915",
]
